requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.2.0
opencv-python==4.9.0.80
transformers==4.41.1
torch>=2.0.0
numpy==1.26.4
scipy==1.13.1
easyocr==1.7.1
gtts==2.5.1
playsound==1.2.2
requests==2.32.2
